+++ CHANGELOG.md
@@ -7,6 +7,9 @@ vNext
 ### Enhancements
 * Support query on lists of realm objects
 
+### Fixed
+* Iterating RealmResult will snapshot the collection if the items are RealmObject from Realm.
+
 
 0.2.0+alpha Release notes (2022-01-31)
 ==============================================================